src/mcqgenerator/MCQGenerator.py
import os
import json
import traceback
import pandas as pd
from dotenv import load_dotenv
from src.mcqgenerator.utils import read_file,get_table_data
from src.mcqgenerator.logger import logging
#imporing necessary packages packages from langchain
from langchain import huggingface_hub
from langchain.prompts import Promptlemplate
from langchain.chains import LLMChain
from langchain.chains import SequentialChain

