requirements.txt
huggingface_hub
transformers
langchain
streamlit
python-dotenv
PyPDF2


-e .